realizing I need to start actually committing thins instead of just playing around
