+++ element_parsers/resp3_simple_string.py
@@ -1,16 +1,17 @@
 from ..util import slice_first_byte
 
 
-def parse_simple_string(data: bytes):
+def parse_simple_string(data: list[str]):
     """ Take a RESP3 simple string representation as bytes and return the 
         string it contains, as a string
     
         example: +OK\r\n
     """
-    _PREFIX = b"+"
-    if slice_first_byte(data) != _PREFIX:
+    _PREFIX = "+"
+    if data[0][0] != _PREFIX:
         raise ValueError(f"Expected '{_PREFIX}' for simple string prefix, got {data[0]}")
-    print("Simple String")
+    data[0] = data[0][1:]
+    return data
 
 
 # def test_simple_string():
